ref: move tree-stratum util functions for general utils
