+++ tests/forestry/preprocessing_tests/age_supplement_test.py
@@ -53,56 +53,6 @@ age_stratums = create_test_stratums(age_stratum_inputs)
 
 
 class TestAgeSupplementing(unittest.TestCase):
-    def test_generate_diameter_threshold(self):
-        assertions = [
-            ((10.0, 20.0), 13.33333),
-            ((5.0, 10.0), 6.66667),
-            ((0.0, 10.0), 0.0),
-            ((5.0, 0.0), 0.0)
-        ]
-
-        for i in assertions:
-            result = round(age_sup.generate_diameter_threshold(i[0][0], i[0][1]), 5)
-            self.assertEqual(result, i[1])
-
-    def test_override_from_diameter(self):
-        initial_stratum = TreeStratum()
-        initial_stratum.mean_diameter = 10.0
-        current_stratum = TreeStratum()
-        current_stratum.mean_diameter = 20.0
-        assertions = [
-            (13.0, current_stratum),
-            (15.0, initial_stratum),
-        ]
-        for i in assertions:
-            reference_tree = ReferenceTree()
-            reference_tree.breast_height_diameter = i[0]
-            result = age_sup.override_from_diameter(initial_stratum, current_stratum, reference_tree)
-            self.assertEqual(i[1], result)
-
-    def test_solve_stratum_supplement(self):
-        reference_tree = ReferenceTree()
-        reference_tree.species = 2
-        reference_tree.breast_height_diameter = 13.0
-        species_diameter = [
-            (1, 3.0),
-            (1, 12.0),
-            (2, 10.0),
-            (2, 5.0),
-            (3, 7.0),
-            (3, 10.0),
-            (3, 5.80)
-        ]
-
-        stratums = []
-        for spe_dia in species_diameter:
-            ts = TreeStratum()
-            ts.species = spe_dia[0]
-            ts.mean_diameter = spe_dia[1]
-            stratums.append(ts)
-        result = age_sup.solve_stratum_supplement(reference_tree, stratums)
-        self.assertEqual(stratums[2], result)
-
     def test_final_tree_strategy(self):
         local_no_age_tree_inputs = [
             Input('001-001-01-3-03-tree', 3, 10.0, None, None, None),
